brain/requirements.txt
httpx
rich
python-dotenv

